services/astra-orchestrator/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
redis>=5.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
boto3>=1.34.0
pybreaker==1.0.0
tenacity>=8.2.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
Pillow>=10.0.0
